Trim down buildSrc Util, rename to JGitUtil
